fixed typeerrors
